v0.4 def update

+++ backend/foodgram/api/serializers.py
@@ -173,10 +173,16 @@ class RecipeSerializer(serializers.ModelSerializer):
 
     def update(self, recipe, validated_data):
         ingredients = validated_data.pop('ingredients')
-        tags = validated_data.pop('tags')
-        IngredientInRecipe.objects.filter(recipe=recipe).delete()
-        self.add_ingredients(ingredients, recipe)
-        recipe.tags.set(tags)
+        if 'ingredients' in self.validated_data:
+            IngredientInRecipe.objects.filter(recipe=recipe).delete()
+            for ingredient in ingredients:
+                IngredientInRecipe.objects.create(
+                    recipe=recipe,
+                    ingredient=ingredient.get('id'),
+                    amount=ingredient.get('amount')
+                )
+        if 'tags' in self.validated_data:
+            recipe.tags.set(validated_data.pop('tags'))
         return super().update(recipe, validated_data)
 
     def to_representation(self, instance):
